moved `OclcNumber` class to `models.py`
